Memperbaiki length integer Mahasiswa

+++ database/migrations/2020_12_31_062052_create_mahasiswas_table.php
@@ -15,7 +15,7 @@ class CreateMahasiswasTable extends Migration
     {
         Schema::create('mahasiswas', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('nim');
+            $table->bigInteger('nim');
             $table->string('nama');
             $table->string('tempat_lahir');
             $table->date('tanggal_lahir');
